requirementos.txt
EasyProcess==1.1
entrypoint2==1.1
evdev==1.6.1
jeepney==0.8.0
MouseInfo==0.1.3
mss==7.0.1
Pillow==9.4.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.0
PyScreeze==0.1.28
python-xlib==0.33
python3-xlib==0.15
pytweening==1.0.4
six==1.16.0
